requirements.txt
docx
pyttsx3-2.90